Create Modal component that will display current items in Cart
Modal component is reusable. This app will most probably use it solely to display items in Cart but making Modal reusable shows good foresight as it is still probable that the Modal component will be used in a different context in the future when new features get added on the app.