Fix bug in Modal component
Bug: When user presses the Esc key on his/her keyboard, the Cart and Checkout
Modals could not be opened again because closing the Modal using Esc key does not update the progress context. User progress context only gets updated when the Close button inside the Modal is pressed to dismiss the Modal. The fix is to call back the progress update function when Modal closes: onClose={callbackFxn}.